refactor a bit to fix circular dependency spotted by oxlint
